requirements.txt
streamlit==1.38.0
pillow==10.3.0
numpy==1.26.4
matplotlib==3.8.4
typing_extensions==4.9.0
torch>=2.2
torchvision>=0.17
pillow>=9.5
# KHÔNG ghi torch/torchvision ở đây (đã cài ở Dockerfile với +cu121)
